Implemented PDF_DIRFROMMOMENTUM and DPF_FLAT
Also:
- Fixed CleanUpParticle only marking the particle as destroyed if OnParticleDeath returned false
- Fixed an issue where fading would only set the particle as translucent based on the *default* render style, rather than the particle style
- Made HWSprite particleflags 32 bit, so it wouldn't eat any flags over 16

+++ src/rendering/hwrenderer/scene/hw_drawstructs.h
@@ -397,7 +397,7 @@ public:
 
 	bool isparticle = false;
 	bool particlehastexture = false;
-	uint16_t particleflags = 0;
+	uint32_t particleflags = 0;
 	subsector_t* particlesubsector = nullptr;
 
 	bool nomipmap; // force the sprite to have no mipmaps (ensures tiny sprites in the distance stay crisp)
